backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# MongoDB
pymongo>=4.6.0

# ChromaDB
chromadb>=0.4.22

# DeepSeek / LangChain
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
openai>=1.10.0

# JWT
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 文件处理
python-multipart>=0.0.6
PyMuPDF>=1.23.0
python-docx>=1.1.0

# 向量嵌入
sentence-transformers>=2.2.0
scikit-learn>=1.4.0
numpy>=1.26.0

# 其他
python-dotenv>=1.0.0
